Support address bound and outgoing packet NIC
...in routes.

When forwarding is enabled, the source address of a packet may be bound
to a different interface than the outgoing interface. This change
updates stack.Route to hold both NICs so that one can be used to write
packets while the other is used to check if the route's bound address
is valid.

PiperOrigin-RevId: 335144334